+++ android/src/main/java/com/cloudwebrtc/webrtc/utils/ConstraintsArray.java
@@ -39,6 +39,10 @@ public class ConstraintsArray {
         return (String) mArray.get(index);
     }
 
+    public Byte[] getByte(int index){
+        return (Byte[]) mArray.get(index);
+    }
+
     public ConstraintsArray getArray(int index){
         return new ConstraintsArray((ArrayList<Object>)mArray.get(index));
     }
@@ -64,6 +68,8 @@ public class ConstraintsArray {
             return ObjectType.Array;
         } else if (object instanceof Map) {
             return ObjectType.Map;
+        } else if (object instanceof Byte) {
+            return ObjectType.Byte;
         }
         return ObjectType.Null;
     }
@@ -94,6 +100,10 @@ public class ConstraintsArray {
 
     public void pushArray(ConstraintsArray array){
         mArray.add(array.toArrayList());
+    }
+
+    public void pushByte(byte[] value){
+        mArray.add(value);
     }
 
     public void pushMap(ConstraintsMap map){