+++ ios/Classes/FlutterRTCPeerConnection.m
@@ -483,10 +483,11 @@
     peerConnection.dataChannels[dataChannelId] = dataChannel;
     
     FlutterEventChannel *eventChannel = [FlutterEventChannel
-                                         eventChannelWithName:[NSString stringWithFormat:@"cloudwebrtc.com/WebRTC/dataChannelEvent%d", dataChannel.channelId]
+                                         eventChannelWithName:[NSString stringWithFormat:@"FlutterWebRTC/dataChannelEvent%d", dataChannel.channelId]
                                          binaryMessenger:self.messenger];
     
     dataChannel.eventChannel = eventChannel;
+    dataChannel.flutterChannelId = dataChannelId;
     [eventChannel setStreamHandler:dataChannel];
     
     FlutterEventSink eventSink = peerConnection.eventSink;